Separate rustdoc passes into separate modules.
